:lock: Changed code to six digits
Also completed the dashboard for when a user is "unclaimed" (hopefully)

Added a little bit of logic to display when a user is "unclaimed" in the admin page. Displays as "unregistered". Currently admins have no way of getting the code that was generated to send to users. This is obviously an important functionality.

Sped up the animation for custom clickblockers.