:bug: Fixed bug that prevented users from seeing the login page
updated a lot of background colors to be using the theme, also added some colors to the theme to accommodate the extra background colors

Admin users simply cannot add users without triggering an error. I do not understand since it seems to be working fine, throwing an error for no reason I can find.
Temporarily removed the line that would update the company ID for the user since that can be handled with the back end functions. Perhaps updateDoc will simply not work?