:boom: Lots of changes | error handling | cloud function | calendar refreshing | blocker scrolling | display table skeleton
## Added rudimentary error handling
Pretty much everything that has a .then now also has a .catch and will alert the user there was an error and inform them they should reload. Definitely need more in depth error handling for each issue as well as automated reporting for when a user encounters an error

## Cloud functions
Added a function for removing an employee's company from their collection. This was not possible on the front end because of prevent users from editing the Administrative Details document.

*This function has not been tested or deployed*

## Calendar Refreshing
Calendar date array (the selectable dates displayed) should now update when the window is refocused. This means when a user logs in on Monday, closes the app, then reopens the app on Tuesday, Tuesday should now be selectable.

Needs actual testing. Also, it seems that hours worked this week (displayed in HourAdder) doesn't update sometimes when transitioning to the next week. Needs more testing

## ClickBlocker
Added scrolling to the modal for when custom content is too big. Interestingly, when I added two HourAdders on top of each other for testing, it cut off the top third (?) of the first one. I looked for what styling to blame and never found it.

## Display Table Skeleton
Added a skeleton for the admin display table while the company data is loading