:boom: Mostly working!
Loading status is working properly, but due to issues with it being handled by the ConnectionHandler, it is currently being handled by the EmployeeLine instead. This means right now the status is for the table is being handled by the ConnectionHandler, but the status for the empLine is being handled by itself. Thhis needs to be unified so that they are consistent.

Also! Added the Dillahunty Farms logo! Only in the admin page right now, but it looks great! Very happy with it.

TODO:
- Turn that logo thing into something standardized so that a company's admin can upload/customize what they see there, and eventually add styling for what their employees see (not to the extent of MyChart though lol)